Update to message event
Updated the way commands are being shown in the console. Much easier to
get a quick overview of the bot is being used, who is using it, and
where they are using it from.

All of this information is important to know where this bot is heading.
And to improve on the ways people are interacting with it.

I also fixed a small bug where commands would not be recognized if they
were typed with a capital letter. Capitalazation should not matter
anymore.