Widgets BETA
Widgets Beta feature is finally here!

It's very unstable. And there is no checks to make sure the widget
params are properly set.  So better not fuck it up while adding it. :P

More feature and better checks, and just overall improvements will come
in later commits.

For now, just enjoy the simplicity for what it is.

Features:
+ You can add & remove widgets via a command
+ You can add more widgets like you would add a command. I'd suggest
using the time.js as a template for new commands.
+ The minimum time between widget updates are 1 minute. This isn't
real-time down to the second. There has to be some delay.
+ The widgets will be reloaded every time it's called. To make sure it
executes the latest command. This means you can update your widgets on
the fly without restarting the bot. Or even give it a command to reload
the widgets.

Known bugs:
- I had a small issue with check, I had two different check widgets. But
they got merged into one. As in, the info from one of the servers got
displayed in both.

The fix for this, is to use only one check widget. But you can add more
check widgets to the same server if you want to.