Moved my workspace to Visual Studio Code
And thus, I added ESLint.

There is no reason for me to include my workspace, which is local to me,
and my ESLint settings file, which is also local to me.

So they will be ignored by Github from now on.

If you wish to use Visual Studio Code, you may see fit to generate these
files yourself. And there is no need to push them to me either.

Henche the gitignore entries.

Updated the perms for the translate command

Now anyone can use it!
Hooray! :D

Fixed the guild default channel issue

The bot used to crash every time
someone left or joined one of the
guilds the bot was a part of.

Now,
that should no longer happen.

By default, the bot will report whenever
someone leaves / joins on all
guilds it's a part of, to the default
channel on the guild.

Later, when I add proper multi-guild support,
this will be
customizeable, and even able to be turned off.