Fixed the join / leave messages issues
The bot would crash whenever someone left or joined a guild it was a
part of, because it coulden't find the default channel.

Now that is no longer the case.
Added a method for locating the default channel, or try to at least.

Right now, the bot will try to post to any guild it's a part of.

Later on, when I add proper multi-guild support, this will be a
toggleable option, along with what message to display.