Small change to join/leave messages
There's a Discord bug where if you tag a member, that tag name may
display just the ID in a embed. Especially if the user has left the
channel.

And on phones, it won't display the name at all. Only the ID.

With this change, it will display the name with a tag as plain text in a
prenthesis next to the tagged name. So if the tag disapears, you can
still see who it was.