Added some basic sanitization to the widget command
Apparently, I encounrtered a bug where there was a space infront of the
data field. Making the hoostname have a space in it. Which results in an
error in the check command.

Adding some trim() to the input field, and to the check command itself
will make sure this doesn't happen again in the future.