Quick fix to logger
Someone found the public webhook url in logger. And used it to spam messages.

While this only affected the console log on my end. It was still mildly annoying.

And now the webhook url is included in the private config file.

(Which, admitedly, is what it should have been in the first place.)

Anyway, all good now.